+++ .circleci/config.yml
@@ -1,42 +1,92 @@
 version: 2
+
+yarn_anchors:
+  yarn_restore_cache: &yarn_restore_cache
+    restore_cache:
+      keys:
+        - v1-dependencies-{{ arch }}-{{ checksum "yarn.lock" }}
+        - v1-dependencies-{{ arch }}-
+  npm_install_yarn: &npm_install_yarn
+    run:
+      name: npm install yarn
+      command: |
+        YARN_VERSION=1.12.1
+        if [[ $(yarn --version) != $YARN_VERSION ]]; then
+          sudo npm install -g yarn@$YARN_VERSION;
+          sudo chmod a+x /usr/local/bin/yarn
+        fi
+  yarn_install: &yarn_install
+    run: yarn install
+  yarn_save_cache: &yarn_save_cache
+    save_cache:
+      key: v1-dependencies-{{ arch }}-{{ checksum "yarn.lock" }}
+      paths:
+        - ./node_modules
+
 jobs:
-  build:
+  test:
     docker:
-      - image: circleci/node:8.12.0
-
+      - image: circleci/node:10.15.0
     working_directory: ~/repo
-
     steps:
       - checkout
-
-      - restore_cache:
-          keys:
-          - v1-dependencies-{{ arch }}-{{ checksum "yarn.lock" }}
-          - v1-dependencies-{{ arch }}-
-
-      - run:
-          name: npm install yarn
-          command: |
-            YARN_VERSION=1.9.4
-            if [[ $(yarn --version) != $YARN_VERSION ]]; then
-              sudo npm install -g yarn@$YARN_VERSION;
-              sudo chmod a+x /usr/local/bin/yarn
-            fi
-
-      - run: yarn install
-
-      - save_cache:
-          key: v1-dependencies-{{ arch }}-{{ checksum "yarn.lock" }}
-          paths:
-            - ./node_modules
-
-      - run: yarn lint
-
+      - *yarn_restore_cache
+      - *npm_install_yarn
+      - *yarn_install
+      - *yarn_save_cache
       - run:
           name: yarn test
           command: yarn test:coverage --ci --color --reporters=default --reporters=jest-junit --runInBand
           environment:
             JEST_JUNIT_OUTPUT: ./tmp/test-results/jest.xml
-
       - store_test_results:
           path: ./tmp/test-results
+
+  lint:
+    docker:
+      - image: circleci/node:10.15.0
+    working_directory: ~/repo
+    steps:
+      - checkout
+      - *yarn_restore_cache
+      - *npm_install_yarn
+      - *yarn_install
+      - *yarn_save_cache
+      - run: yarn lint
+
+  audit:
+    docker:
+      - image: circleci/node:10.15.0
+    working_directory: ~/repo
+    steps:
+      - checkout
+      - *yarn_restore_cache
+      - *npm_install_yarn
+      - *yarn_install
+      - *yarn_save_cache
+      - run: yarn audit
+
+workflows:
+  version: 2
+  commit:
+    jobs:
+      - test
+      - lint
+      - audit:
+          filters:
+            branches:
+              only:
+                - development
+                - master
+  cron:
+    jobs:
+      - test
+      - lint
+      - audit
+    triggers:
+      - schedule:
+          cron: "0 13 * * *"
+          filters:
+            branches:
+              only:
+                - master
