+++ .eslintignore
@@ -1,2 +1,1 @@
-package.json
 src/serviceWorker.ts
