+++ .eslintrc.json
@@ -1,5 +1,6 @@
 {
-  "plugins": ["jest", "prettier"],
+  "parser": "@typescript-eslint/parser",
+  "plugins": ["@typescript-eslint", "jest", "prettier"],
   "rules": {
     "no-unused-vars": [
       "error",
@@ -10,10 +11,21 @@
       }
     ],
     "no-var": "error",
-    "prettier/prettier": "error"
+    "prettier/prettier": "error",
+    "@typescript-eslint/explicit-function-return-type": [
+      "error",
+      {
+        "allowTypedFunctionExpressions": true,
+        "allowExpressions": true
+      }
+    ]
   },
   "extends": [
+    "react-app",
     "eslint:recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:jest/recommended",
+    "prettier/@typescript-eslint",
     "prettier"
   ]
 }
