+++ .gitignore
@@ -1,16 +1,14 @@
-# See https://help.github.com/ignore-files/ for more about ignoring files.
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
 /node_modules
+yarn.lock
 
 # testing
 /coverage
 
 # production
 /build
-
-# compiled css
-*.css
 
 # misc
 .DS_Store
