+++ .tool-versions
@@ -1,1 +1,1 @@
-nodejs 8.12.0
+nodejs 10.15.0
