--- images.d.ts
@@ -1,3 +1,0 @@
-declare module "*.svg";
-declare module "*.png";
-declare module "*.jpg";
