+++ package.json
@@ -1,76 +1,82 @@
 {
-  "name": "app-prototype",
+  "name": "npx",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@types/jest": "^23.3.2",
-    "@types/react": "^16.4.14",
-    "@types/react-dom": "^16.0.7",
-    "babel-eslint": "^7.2.3",
+    "@types/enzyme": "^3.10.3",
+    "@types/enzyme-adapter-react-16": "^1.0.5",
+    "@types/jest": "24.0.15",
+    "@types/node": "12.6.8",
+    "@types/react": "16.8.23",
+    "@types/react-dom": "16.8.4",
     "classnames": "^2.2.6",
-    "eslint": "^4.19.1",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-config-react-app": "^2.1.0",
-    "eslint-plugin-flowtype": "^2.34.1",
-    "eslint-plugin-import": "^2.6.0",
-    "eslint-plugin-jest": "^21.18.0",
-    "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-prettier": "^2.6.1",
-    "eslint-plugin-react": "^7.1.0",
-    "husky": "^0.14.3",
-    "jest-junit": "^5.1.0",
-    "lint-staged": "^7.2.0",
-    "node-sass-chokidar": "^1.3.3",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.14.0",
+    "eslint-config-prettier": "^6.0.0",
+    "eslint-plugin-jest": "^22.6.4",
+    "eslint-plugin-prettier": "^3.1.0",
+    "husky": "^2.4.0",
+    "jest-enzyme": "^7.0.2",
+    "jest-junit": "^6.4.0",
+    "lint-staged": "^8.2.0",
+    "node-sass-chokidar": "^1.3.5",
     "npm-run-all": "^4.1.3",
-    "prettier": "^1.13.7",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-scripts-ts": "^2.17.0",
-    "serve": "^9.4.0",
-    "stylelint": "^9.4.0",
-    "stylelint-config-recommended-scss": "^3.2.0",
-    "stylelint-config-standard": "^18.2.0",
+    "prettier": "1.18.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-scripts": "3.0.1",
+    "serve": "^11.0.1",
+    "shelljs": "^0.8.2",
+    "stylelint": "^10.0.1",
+    "stylelint-config-recommended-scss": "^3.3.0",
+    "stylelint-config-standard": "^18.3.0",
     "stylelint-declaration-use-variable": "^1.6.1",
-    "stylelint-scss": "^3.2.0",
-    "tslint": "^5.11.0",
-    "tslint-config-prettier": "^1.15.0",
-    "tslint-react": "^3.6.0",
-    "typescript": "^3.0.3"
+    "stylelint-scss": "^3.8.0",
+    "typescript": "^3.5.3"
   },
   "scripts": {
-    "start": "npm-run-all -p start:css start:ts",
-    "start:css": "npm-run-all build:css build:css:watch",
-    "start:ts": "react-scripts-ts start",
-    "build": "npm-run-all build:css build:ts",
-    "build:css": "node-sass-chokidar src/ -o src/",
-    "build:css:watch": "node-sass-chokidar src/ -o src/ --watch --recursive",
-    "build:ts": "react-scripts-ts build",
+    "prestart": "node bin/pre-start",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
     "lint": "npm-run-all -c lint:*",
-    "lint:eslint": "eslint 'src/**/*.{js,jsx}'",
-    "lint:prettier": "prettier --write **/*.{js,jsx,ts,tsx,css,scss,json,md}",
+    "lint:eslint": "eslint 'src/**/*.{ts,tsx}'",
     "lint:stylelint": "stylelint src/**/*.scss",
-    "lint:tslint": "tslint 'src/**/*.{ts,tsx}'",
-    "pretest": "npm-run-all build:css",
-    "test": "react-scripts-ts test --env=jsdom",
-    "pretest:coverage": "npm-run-all build:css",
-    "test:coverage": "react-scripts-ts test --env=jsdom --coverage",
-    "eject": "react-scripts-ts eject",
-    "precommit": "lint-staged",
+    "lint:prettier": "prettier --write **/*.{ts,tsx,json,scss,md}",
+    "test": "react-scripts test",
+    "test:coverage": "react-scripts test --env=jsdom --coverage",
+    "eject": "react-scripts eject",
     "heroku-postbuild": "npm-run-all build"
   },
   "lint-staged": {
     "src/**/*.{js,jsx}": "eslint",
-    "src/**/*.{ts,tsx}": "tslint",
     "src/**/*.scss": "stylelint"
   },
   "engines": {
-    "node": "8.12.0",
-    "yarn": "1.9.4"
+    "node": "10.15.0",
+    "yarn": ">=1.12.1"
   },
   "jest": {
     "collectCoverageFrom": [
       "src/**/*.{ts,tsx}",
-      "!src/registerServiceWorker.ts"
+      "!src/serviceWorker.ts"
     ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not ie <= 11",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
   }
 }
