+++ src/App.scss
@@ -1,28 +1,34 @@
-$gray-22: #222;
-$white: white;
+@import "./theme/colors.scss";
+@import "./theme/typography.scss";
 
 .App {
   text-align: center;
+}
 
-  &__logo {
-    animation: App-logo-spin infinite 20s linear;
-    height: 80px;
-  }
+.App__logo {
+  animation: App-logo-spin infinite 20s linear;
+  height: 40vmin;
+  pointer-events: none;
+}
 
-  &__header {
-    background-color: $gray-22;
-    height: 150px;
-    padding: 20px;
-    color: $white;
-  }
+.App__header {
+  @include headline-text;
 
-  &__title {
-    font-size: 1.5em;
-  }
+  background-color: $gray-28;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  color: $white;
+}
 
-  &__intro {
-    font-size: large;
-  }
+.App__code {
+  font-family: $font-monospace;
+}
+
+.App__link {
+  color: $blue;
 }
 
 @keyframes App-logo-spin {
