+++ src/index.tsx
@@ -1,8 +1,12 @@
-import * as React from "react";
-import * as ReactDOM from "react-dom";
-import "./index.css";
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.scss";
 import App from "./App";
-import registerServiceWorker from "./registerServiceWorker";
+import * as serviceWorker from "./serviceWorker";
 
 ReactDOM.render(<App />, document.getElementById("root"));
-registerServiceWorker();
+
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: https://bit.ly/CRA-PWA
+serviceWorker.unregister();
