--- tslint.json
@@ -1,10 +1,0 @@
-{
-  "rules": {
-    "member-access": [true, "no-public"],
-    "ordered-imports": false
-  },
-  "extends": ["tslint-react", "tslint:recommended", "tslint-config-prettier"],
-  "linterOptions": {
-    "exclude": ["node_modules/**/*.ts", "coverage/lcov-report/*.js"]
-  }
-}
